approx structure for kafka test utils
