+++ Sources/KafkaTestUtils/KafkaTestLogger.swift
@@ -1,0 +1,23 @@
+//===----------------------------------------------------------------------===//
+//
+// This source file is part of the swift-kafka-client open source project
+//
+// Copyright (c) 2023 Apple Inc. and the swift-kafka-client project authors
+// Licensed under Apache License v2.0
+//
+// See LICENSE.txt for license information
+// See CONTRIBUTORS.txt for the list of swift-kafka-client project authors
+//
+// SPDX-License-Identifier: Apache-2.0
+//
+//===----------------------------------------------------------------------===//
+
+import Logging
+
+extension Logger {
+    public static var kafkaTest: Logger {
+        var logger = Logger(label: "kafka.test")
+        logger.logLevel = .info
+        return logger
+    }
+}
